fixup! fixup! only read log configuration files ending with .yml or .yaml
